test(e2e): fix html cases occasionally failed
