+++ cspell.json
@@ -28,6 +28,7 @@
     "ontext",
     "osascript",
     "outro",
+    "pageerror",
     "pluggable",
     "pmmmwh",
     "rebranded",
