+++ e2e/cases/html/combined/src/index.js
