+++ e2e/cases/html/combined/static/bar.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>custom template</title>
+    <script>window.type = '<%= type %>'</script>
+  </head>
+  <body>
+    <div id="test-template">bar</div>
+    <div id="root" />
+  </body>
+</html>
