+++ e2e/cases/html/combined/static/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>custom template</title>
+    <script>window.foo = '<%= foo %>'</script>
+  </head>
+  <body>
+    <div id="test-template">xxx</div>
+    <div id="root" />
+  </body>
+</html>
