+++ e2e/cases/html/combined/tsconfig.json
