+++ e2e/cases/html/mount-id/index.test.ts
@@ -1,0 +1,48 @@
+import { join } from 'path';
+import { expect, test } from '@playwright/test';
+import { fse } from '@rsbuild/shared';
+import { build, getHrefByEntryName } from '@scripts/shared';
+
+test.describe('should render mountId correctly', () => {
+  let rsbuild: Awaited<ReturnType<typeof build>>;
+
+  test.beforeAll(async () => {
+    rsbuild = await build({
+      cwd: __dirname,
+      runServer: true,
+      rsbuildConfig: {
+        html: {
+          mountId: 'app',
+        },
+      },
+    });
+  });
+
+  test.afterAll(async () => {
+    await rsbuild.close();
+  });
+
+  test('mountId', async ({ page }) => {
+    const errors: string[] = [];
+    page.on('pageerror', (err) => errors.push(err.message));
+
+    await page.goto(getHrefByEntryName('index', rsbuild.port));
+
+    const test = page.locator('#test');
+    await expect(test).toHaveText('Hello Rsbuild!');
+
+    expect(errors).toEqual([]);
+  });
+
+  test('inject default (head)', async () => {
+    const pagePath = join(rsbuild.distPath, 'index.html');
+    const content = await fse.readFile(pagePath, 'utf-8');
+
+    expect(
+      /<head>[\s\S]*<script[\s\S]*>[\s\S]*<\/head>/.test(content),
+    ).toBeTruthy();
+    expect(
+      /<body>[\s\S]*<script[\s\S]*>[\s\S]*<\/body>/.test(content),
+    ).toBeFalsy();
+  });
+});
