+++ e2e/cases/html/output-structure/src/index.js
@@ -1,0 +1,1 @@
+console.log('1');
