+++ e2e/cases/html/template/index.test.ts
@@ -1,6 +1,6 @@
 import path from 'path';
 import { expect, test } from '@playwright/test';
-import { build } from '@scripts/shared';
+import { build, getHrefByEntryName } from '@scripts/shared';
 
 test('should set template via function correctly', async () => {
   const rsbuild = await build({
@@ -8,7 +8,7 @@ test('should set template via function correctly', async () => {
     rsbuildConfig: {
       source: {
         entry: {
-          index: path.resolve(__dirname, './src/index.ts'),
+          index: path.resolve(__dirname, './src/index.js'),
           foo: path.resolve(__dirname, './src/foo.js'),
         },
       },
@@ -34,4 +34,31 @@ test('should set template via function correctly', async () => {
   const indexHtml =
     files[Object.keys(files).find((file) => file.endsWith('index.html'))!];
   expect(indexHtml).toContain('<div id="test-template">xxx</div>');
+});
+
+test('should allow to access templateParameters', async ({ page }) => {
+  const rsbuild = await build({
+    cwd: __dirname,
+    runServer: true,
+    rsbuildConfig: {
+      html: {
+        template: './static/index.html',
+        templateParameters: {
+          foo: 'bar',
+        },
+      },
+    },
+  });
+
+  await page.goto(getHrefByEntryName('index', rsbuild.port));
+
+  const testTemplate = page.locator('#test-template');
+  await expect(testTemplate).toHaveText('xxx');
+
+  const testEl = page.locator('#test');
+  await expect(testEl).toHaveText('Hello Rsbuild!');
+
+  await expect(page.evaluate(`window.foo`)).resolves.toBe('bar');
+
+  await rsbuild.close();
 });
